requirements.txt
# Core dependencies (used by src/)
fastapi==0.121.2
uvicorn==0.38.0
python-multipart==0.0.21
pydantic==2.12.4
pydantic-settings==2.7.1

# LangChain ecosystem
langchain-core>=0.3.17
langchain-text-splitters==0.3.2
langchain==0.3.7
langchain-google-genai==2.0.5

# LangGraph for agents
langgraph==0.2.55

# Vector database
weaviate-client==4.10.2

# Document processing
pdfminer.six==20231228
unstructured[pdf]==0.16.12
python-magic==0.4.27

# Required transitive dependencies
numpy==1.26.4
protobuf==5.29.5
requests==2.32.5
tqdm==4.67.1
